# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CarryCaro API
  version: 1.0.0
  description: API documentation for CarryCaro authentication, profile, cities, countries, and trips.
servers:
  - url: /api
    description: Current API server
...
paths:
...
  /auth/social-login:
    post:
      tags: [Auth]
      summary: Log in or register with a social provider
      description: Accepts a provider identity from a trusted mobile or frontend social-auth flow.
      operationId: socialLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialLoginRequest'
      responses:
        '200':
          description: Social login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Sanctum token
...
  responses:
...
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        message:
          type: string
          nullable: true
          example: Registration successful. Please verify your email.
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          example: 1|plain-text-token
...
    SocialLoginRequest:
      type: object
      required: [provider, provider_id, name, email]
      properties:
        provider:
          type: string
          enum: [google]
          example: google
        provider_id:
          type: string
          example: '10987654321'
        name:
          type: string
          maxLength: 255
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
        profile_image:
          type: string
          nullable: true
          example: https://example.com/avatar.jpg
...
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
        phone_number:
          type: string
          nullable: true
          example: '+491234567890'
        facebook_profile:
          type: string
          nullable: true
          example: https://facebook.com/jane
        profile_image:
          type: string
          nullable: true
          example: https://example.com/avatar.jpg
        timezone:
          type: string
          nullable: true
          example: Europe/Berlin
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
...
    ValidationErrorResponse:
      type: object
      required: [message, errors]
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - The email field is required.